+++ src/main/java/ru/dm/shop/service/impl/CartProductServiceImpl.java
@@ -17,7 +17,6 @@ import ru.dm.shop.service.ProductService;
 import ru.dm.shop.service.UserService;
 
 import javax.annotation.Resource;
-import java.util.List;
 
 @Service
 public class CartProductServiceImpl implements CartProductService {
@@ -38,7 +37,7 @@ public class CartProductServiceImpl implements CartProductService {
         cartProduct = cartProductRepository.findByUserIdAndProductId(userService.findByUsername(authentication.getName()).getId(), productService.findByArticule(product.getArticule()).getId());
 
         if (cartProduct != null) {
-            int count = cartProduct.getCount();
+            Long count = cartProduct.getCount();
             product = productService.findByArticule(product.getArticule());
             cartProductRepository.updateProductInCart(count + 1, userService.findByUsername(authentication.getName()).getId(), product);
         }
@@ -61,7 +60,7 @@ public class CartProductServiceImpl implements CartProductService {
         CartProduct cartProduct;
         cartProduct = cartProductRepository.findByUserIdAndProductId(userService.findByUsername(authentication.getName()).getId(), productService.findByArticule(product.getArticule()).getId());
 
-        int count = cartProduct.getCount();
+        Long count = cartProduct.getCount();
         product = productService.findByArticule(product.getArticule());
         cartProductRepository.updateProductInCart(amount, userService.findByUsername(authentication.getName()).getId(), product);
 
@@ -87,7 +86,7 @@ public class CartProductServiceImpl implements CartProductService {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         CartProduct cartProduct = cartProductRepository.findByUserIdAndProductId(userService.findByUsername(authentication.getName()).getId(), productService.findByArticule(product.getArticule()).getId());
 
-        int count = cartProduct.getCount();
+        Long count = cartProduct.getCount();
         if (count > 1) {
             product = productService.findByArticule(product.getArticule());
             cartProductRepository.updateProductInCart(count - 1, userService.findByUsername(authentication.getName()).getId(), product);
